Put everything under test_generics_649.py
