--- tests/generics.py
@@ -1,14 +1,0 @@
-"""Generic classes for testing."""
-
-from __future__ import annotations
-
-from typing import Generic, TypeVar
-
-from attrs import define
-
-T = TypeVar("T")
-
-
-@define
-class GenericClass(Generic[T]):
-    t: T
